+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/SoundExporter.java
@@ -147,7 +147,7 @@ public class SoundExporter {
                 flv.writeHeader(true, false);
                 List<SoundStreamBlockTag> blocks = sh.getBlocks();
 
-                int ms = (int) (1000.0f / ((float) ((Tag) st).getSwf().frameRate));
+                int ms = (int) (1000.0 / ((Tag) st).getSwf().frameRate);
                 for (int b = 0; b < blocks.size(); b++) {
                     byte[] data = blocks.get(b).streamSoundData.getRangeData();
                     if (st.getSoundFormatId() == 2) { //MP3
